+++ src/DBFTPlugin/ConsensusMessageType.cs
@@ -4,8 +4,12 @@ namespace Neo.Consensus
     {
         ChangeView = 0x00,
 
+        TXHashesRequest = 0x10,
+        TXHashesResponce = 0x11,
+
         PrepareRequest = 0x20,
         PrepareResponse = 0x21,
+
         Commit = 0x30,
 
         RecoveryRequest = 0x40,
