+++ src/DBFTPlugin/Messages/ChangeView.cs
